% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel - Copy.R, R/sirmodel.R
\name{compartmental_sirmodel}
\alias{compartmental_sirmodel}
\title{Stochastic SIR model, compartmental

This takes N iterations for time step dt}
\usage{
compartmental_sirmodel(end_time, pars = NULL)

compartmental_sirmodel(end_time, pars = NULL)
}
\arguments{
\item{end_time}{en time of data}

\item{pars}{parameter list}
}
\value{
dataframe

dataframe
}
\description{
Stochastic SIR model, compartmental

This takes N iterations for time step dt

Stochastic SIR model, compartmental

This takes N iterations for time step dt
}
\examples{
compartmental_sirmodel(100, list())
run_with_repetitions(100, 1, list(), FALSE)
compartmental_sirmodel(100, list())
run_with_repetitions(100, 1, list(), FALSE)
}
