% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_events}
\alias{get_events}
\title{Events list}
\usage{
get_events(overrides = list())
}
\arguments{
\item{overrides}{use a named events list instead of defaults}
}
\value{
events
}
\description{
Events list
}
