% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Default parameters for SIR model}
\usage{
get_parameters(overrides = list())
}
\arguments{
\item{overrides}{use a named parameter list instead of defaults
Parameters defined below
\itemize{
\item \code{pars} list of parameters
\item \code{beta} contact rate
\item \verb{nu rate} of recovery
\item \verb{mu rate} of death
\item \code{prop_immune} proportion of people who are immune
\item \verb{N number} of people being investigated
\item \code{num} used for countlim/num time points
\item \code{I0} initial number of infected people
\item \code{dt} time step
\item \code{I0_at_steady_state} boolean value
\item \code{plotsave} TRUE if plotted data is to be saved to file
}}
}
\value{
parameter list
}
\description{
Default parameters for SIR model
}
