% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_variables}
\alias{get_variables}
\title{Variable list: immune, age, birth or location}
\usage{
get_variables(overrides = list())
}
\arguments{
\item{overrides}{use a named variable list instead of defaults}
}
\value{
variable list
}
\description{
Variable list: immune, age, birth or location
}
