% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel.R
\name{individual_sirmodel}
\alias{individual_sirmodel}
\title{Stochastic SIR model, individual}
\usage{
individual_sirmodel(end_time, pars = NULL, vars = NULL, events = NULL)
}
\arguments{
\item{end_time}{end time for data}

\item{pars}{parameter list}
}
\value{
dataframe
}
\description{
Takes timestep steps over time for population N
}
\examples{
individual_sirmodel(100, list())
}
