% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel - Copy.R, R/sirmodel.R
\name{render_state_sizes}
\alias{render_state_sizes}
\title{Renders the sizes for S, I, R}
\usage{
render_state_sizes(S, I, R, human)

render_state_sizes(S, I, R, human)
}
\arguments{
\item{S}{S}

\item{I}{I}

\item{R}{R}

\item{human}{human}
}
\description{
Renders the sizes for S, I, R

Renders the sizes for S, I, R
}
\examples{
render_state_sizes(S, I, R, human)

render_state_sizes(S, I, R, human)

}
