% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel.R
\name{run_with_repetitions}
\alias{run_with_repetitions}
\title{Run the simulation with repetitions}
\usage{
run_with_repetitions(end_time, repetitions, pars, parallel = FALSE)
}
\arguments{
\item{end_time}{end time for run}

\item{repetitions}{n times to run the simulation}

\item{pars}{parameter list}

\item{parallel}{execute runs in parallel, TRUE or FALSE}
}
\value{
dataframe
}
\description{
Run the simulation with repetitions
}
