% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel.R
\name{sirmodel}
\alias{sirmodel}
\title{Stochastic SIR model}
\usage{
sirmodel(end_time, pars = NULL)
}
\arguments{
\item{end_time}{end time of data}

\item{pars}{list of parameters}
}
\value{
dataframe
}
\description{
This takes N iterations for time step dt
}
\examples{
sirmodel(100, list())
run_with_repetitions(100, 1, list(), FALSE)
}
