% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel.R
\name{compartmental_sirmodel}
\alias{compartmental_sirmodel}
\title{Stochastic SIR model}
\usage{
compartmental_sirmodel(end_time, pars = NULL)
}
\arguments{
\item{end_time}{en time of data}

\item{pars}{parameter list}
}
\value{
dataframe
}
\description{
This takes N iterations for time step dt
}
\examples{
compartmental_sirmodel(100, list())
run_with_repetitions(100, 1, list(), FALSE)
}
