% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Default parameters for SIR model}
\usage{
get_parameters(overrides = list())
}
\arguments{
\item{overrides}{use a named parameter list instead of defaults
Parameters defined below
\itemize{
\item \code{pars} list of parameters
Compartmental
\item \code{beta} contact rate
\item \verb{nu rate} of recovery
\item \verb{mu rate} of death
\item \code{prop_immune} proportion of people who are immune
\item \verb{N number} of people being investigated
\item \code{num} used for countlim/num time points
\item \code{I0} initial number of infected people
\item \code{dt} time step
\item \code{I0_at_steady_state} boolean value
\item \code{n_deaths_S} number of deaths at susceptible stage
\item \code{n_infections_S} number of infections at susceptible stage
\item \code{n_deaths_I} number of deaths at infected stage
\item \code{n_recoveries_I} number of recoveries at infected stage
\item \code{n_deaths_R} number of deaths at recovered stage
\item \code{n_births} number of births
Individual only
\item \code{variations} any variations to take into account FALSE if not
\item \code{average_age} average age for population
\item \code{includeage} TRUE if age used
\item \code{includebirth} TRUE if immunity used
\item \code{includeimmune} TRUE if immunity used
\item \code{includelocation} TRUE if location used
\item \code{infection_rate} rate of infection
\item \code{recovery_rate} rate of recovery
}}
}
\value{
list
}
\description{
Default parameters for SIR model
}
