% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel.R
\name{individual_I_to_R}
\alias{individual_I_to_R}
\title{Recovery_process -> I to R}
\usage{
individual_I_to_R(I, R, human, immunity, age, location, pars = NULL)
}
\arguments{
\item{I}{infected}

\item{R}{recovered}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Recovery_process -> I to R
}
\examples{
individual_I_to_R(I, R, human, immunity, age, location, pars)
}
