% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel - Copy.R, R/sirmodel.R
\name{individual_R_to_S}
\alias{individual_R_to_S}
\title{Recovered to susceptible -> R to S}
\usage{
individual_R_to_S(S, R, human, immunity, age, location, pars = NULL)

individual_R_to_S(S, R, human, immunity, age, location, pars = NULL)
}
\arguments{
\item{S}{susceptible}

\item{R}{recovered}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Recovered to susceptible -> R to S

Recovered to susceptible -> R to S
}
\examples{
individual_R_to_S(S, R, human, immunity, age, location, pars)
individual_R_to_S(S, R, human, immunity, age, location, pars)
}
