% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirmodel.R
\name{individual_S_to_I}
\alias{individual_S_to_I}
\title{Infection_process -> S to I}
\usage{
individual_S_to_I(S, I, human, pars = NULL)
}
\arguments{
\item{S}{susceptible}

\item{I}{infected}

\item{human}{human}

\item{pars}{parameter list}
}
\description{
Infection_process -> S to I
}
\examples{
individual_S_to_I(S, I, human, pars)
}
